requirements.txt
python-dotenv
phpserialize